make code better:
1.extract server/api.go

2.make webapi reading prom with reusing server's API,not a new prom client
