+++ src/pkg/prom/client_option.go
@@ -1,0 +1,9 @@
+package prom
+
+type ClientOptions struct {
+	BasicAuthUser string
+
+	BasicAuthPass string
+
+	Headers []string
+}
