+++ src/webapi/router/router_prometheus.go
@@ -2,14 +2,14 @@ package router
 
 import (
 	"context"
-	"encoding/json"
-	. "github.com/didi/nightingale/v5/src/pkg/common"
+	. "github.com/didi/nightingale/v5/src/pkg/prom"
 	"github.com/didi/nightingale/v5/src/webapi/reader"
+	"github.com/prometheus/common/model"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
-	"strconv"
 	"strings"
+	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/toolkits/pkg/ginx"
@@ -30,18 +30,19 @@ type batchQueryForm struct {
 }
 
 type batchQueryRes struct {
-	Status string            `json:"status"`
-	Data   []json.RawMessage `json:"data"`
+	Status string        `json:"status"`
+	Data   []model.Value `json:"data"`
 }
 
 func promBatchQueryRange(c *gin.Context) {
-	target, cluster := getClusterAndTarget(c)
+
+	xcluster := c.GetHeader("X-Cluster")
 	var f batchQueryForm
 	err := c.BindJSON(&f)
 	if err != nil {
 		c.String(500, "%s", err.Error())
 	}
-	res, err := batchQueryRange(target, cluster, f.Queries)
+	res, err := batchQueryRange(xcluster, f.Queries)
 
 	ginx.NewRender(c).Data(res, err)
 }
@@ -112,98 +113,31 @@ func clustersGets(c *gin.Context) {
 	ginx.NewRender(c).Data(names, nil)
 }
 
-func batchQueryRange(target *url.URL, cluster *prom.ClusterType, data []queryFormItem) (batchQueryRes, error) {
+func batchQueryRange(cluster string, data []queryFormItem) (batchQueryRes, error) {
+
 	var res batchQueryRes
 
-	req, err := http.NewRequest("GET", target.Path, nil)
-	if err != nil {
-		return res, err
-	}
-	req.URL.Scheme = target.Scheme
-	req.URL.Host = target.Host
-	req.Header.Set("Host", target.Host)
-
-	req.URL.Path = strings.TrimRight(target.Path, "/") + "/api/v1/query_range"
-
-	if target.RawQuery != "" {
-		req.URL.RawQuery = target.RawQuery + "&" + req.URL.RawQuery
-	}
-
-	if _, ok := req.Header["User-Agent"]; !ok {
-		req.Header.Set("User-Agent", "")
-	}
-
-	if cluster.Opts.BasicAuthUser != "" {
-		req.SetBasicAuth(cluster.Opts.BasicAuthUser, cluster.Opts.BasicAuthPass)
-	}
-
-	headerCount := len(cluster.Opts.Headers)
-	if headerCount > 0 && headerCount%2 == 0 {
-		for i := 0; i < len(cluster.Opts.Headers); i += 2 {
-			req.Header.Add(cluster.Opts.Headers[i], cluster.Opts.Headers[i+1])
-			if cluster.Opts.Headers[i] == "Host" {
-				req.Host = cluster.Opts.Headers[i+1]
-			}
-		}
-	}
-
 	for _, item := range data {
-		q := req.URL.Query()
-		q.Set("query", item.Query)
-		q.Set("end", strconv.FormatInt(item.End, 10))
-		q.Set("step", strconv.FormatInt(item.Step, 10))
-		q.Set("start", strconv.FormatInt(item.Start, 10))
-		req.URL.RawQuery = q.Encode()
-		resp, body, err := reader.Client.Do(context.Background(), req)
-
+		client := reader.Reader.Clients[cluster]
+		r := Range{
+			Start: time.Unix(item.Start, 0),
+			End:   time.Unix(item.End, 0),
+			Step:  time.Duration(item.Step) * time.Second,
+		}
+		resp, _, err := client.QueryRange(context.Background(), item.Query, r)
 		if err != nil {
 			return res, err
 		}
-
-		code := resp.StatusCode
-
-		if code/100 != 2 && !ApiError(code) {
-			errorType, errorMsg := ErrorTypeAndMsgFor(resp)
-			return batchQueryRes{}, &Error{
-				Type:   errorType,
-				Msg:    errorMsg,
-				Detail: string(body),
-			}
-		}
-
-		var result ApiResponse
-
-		if http.StatusNoContent != code {
-			if jsonErr := json.Unmarshal(body, &result); jsonErr != nil {
-				return batchQueryRes{}, &Error{
-					Type: ErrBadResponse,
-					Msg:  jsonErr.Error(),
-				}
-			}
-		}
-
-		if ApiError(code) != (result.Status == "error") {
-			err = &Error{
-				Type: ErrBadResponse,
-				Msg:  "inconsistent body for response code",
-			}
-		}
-
-		if ApiError(code) && result.Status == "error" {
-			err = &Error{
-				Type: result.ErrorType,
-				Msg:  result.Error,
-			}
-		}
-
-		res.Data = append(res.Data, result.Data)
+		res.Data = append(res.Data, resp)
 	}
 	res.Status = "success"
 	return res, nil
 }
 
 func getClusterAndTarget(c *gin.Context) (*url.URL, *prom.ClusterType) {
+
 	xcluster := c.GetHeader("X-Cluster")
+
 	if xcluster == "" {
 		c.String(http.StatusBadRequest, "X-Cluster missed")
 		return nil, nil
