+++ src/webapi/webapi.go
@@ -119,7 +119,7 @@ func (a Webapi) initialize() (func(), error) {
 	if err = prom.Init(); err != nil {
 		return nil, err
 	}
-
+	// init reader clients
 	if err = reader.Init(); err != nil {
 		return nil, err
 	}
